requirements.txt
streamlit
requests
beautifulsoup4
serpapi
python-dotenv
